// simulation/types.ts
export type Message = string

export type CheckResult = {
  info: Message[]
  warnings: Message[]
  errors: Message[]
}

export type ProposalData = {
  governor: Contract
  timelock: Contract
  provider: JsonRpcProvider
}

export interface ProposalCheck {
  name: string
  checkProposal(proposal: ProposalEvent, tx: TenderlySimulation, deps: ProposalData): Promise<CheckResult>
}

export interface AllCheckResults {
  [checkId: string]: { name: string; result: CheckResult }
}